Plot function for 'BATSS' outputs

Plot for objects of class 'batss'

Usage

# S3 method for class 'batss'
plot(
  x,
  type = "size",
  hypothesis = "H1",
  title = TRUE,
  legend = TRUE,
  col = c("#008B0040", "#8B3A3A40", "#8B897040", "#FF990075", "blue"),
  ...
)

Arguments

x

An object of class 'batss' (i.e., output of the function batss.glm).

type

A character string indicating the type of plot with options 'size' (default) to display the total and per group sample size observed in the Monte Carlo trials, and 'estimates' to display the Monte Carlo trial target estimates as a function of the sample size.

hypothesis

A character string indicating which alternative hypothesis to use for analyses considering both "H0" and "H1", with options "H1" (default) and "H0".

title

Either a logical indicating if a title should be added or a string (of class character) indicating the title to be added. If title equals TRUE (default), the title 'Under 'H1' or 'Under 'H0' (depending on the argument hypothesis) is added to the outer margin of the plot. No outer margin space is added if title = FALSE.

legend

a logical (with default set to TRUE) indicating if, when 'type = estimates', a legend should be added at the bottom of the plot.

col

a vector of length 5 specifiying the colour respectively assigned to i/ efficacy, ii/ futility, iii/ neither or iv/ both when color-coding trials when type = "estimates". The 5th colour is used for lines. Default to c("#8B897040","#008B0040","#8B3A3A40","#FF990075","blue") where the 2 last digits of the long hexadecimal strings of colours 1 to 4 specify the level of transluency. Refer to the Section 'colour specification' in par for details. If the length of col equals 1, the same colour is used for all cases. When type = "size", the 3rd and 5th colours of the vector col are used to display boxplots and lines.

...

Additional arguments affecting the plot produced, like ylim and ylab.

Value

Generates graphical displays of results for objects of class 'batss'.

See also